+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🐛 Fixed
 
 - **Active Talentbuild:** Wrong default value initilized
+- **Aura Tracker:** Missing sound messages on import of disabled custom sound effects
 
 ---